vi is tired
