+++ green-ev/Dockerfile
@@ -1,11 +1,11 @@
-FROM oven/bun:1
+FROM node:24
 
 WORKDIR /green-ev
 
 COPY package.json .
 COPY bun.lock .
 
-RUN bun install
+RUN npm install --global bun && bun install
 
 EXPOSE 3000
 CMD ["bun", "--bun", "run", "dev"]
