+++ green-ev/app/dashboard/page.tsx
@@ -19,9 +19,10 @@ export default function DashboardPage() {
         const stationsData = user?.role === 'OPERATOR' 
           ? await chargingStationApi.getByOperator()
           : await chargingStationApi.getAll()
-        setStations(stationsData)
+        setStations(stationsData || [])
       } catch (error) {
         console.error('Error fetching stations:', error)
+        setStations([])
       }
     }
 
@@ -29,6 +30,10 @@ export default function DashboardPage() {
   }, [user?.role])
 
   const getAvailabilityColor = (station: StationsSpots) => {
+    if (!station.spots || !Array.isArray(station.spots)) {
+      return 'bg-gray-100 text-gray-800'
+    }
+    
     const freeSpots = station.spots.filter(spot => spot.state === 'FREE').length
     const totalSpots = station.spots.length
     const freePercentage = (freeSpots / totalSpots) * 100
@@ -44,62 +49,69 @@ export default function DashboardPage() {
       
       <div className="grid gap-6 md:grid-cols-2 lg:grid-cols-3">
         {stations.map((station) => (
-          <Card key={station.chargingStation.id} className="overflow-hidden">
-            <div className="relative h-48 w-full bg-gray-100 flex items-center justify-center">
-              <Image
-                src="/charging-station.jpg"
-                alt={station.chargingStation.name}
-                fill
-                className="object-cover"
-                onError={(e) => {
-                  // Hide the image on error
-                  const target = e.target as HTMLImageElement
-                  target.style.display = 'none'
-                }}
-              />
-              <div className="absolute inset-0 flex items-center justify-center">
-                <Zap className="w-24 h-24 text-gray-400" />
-              </div>
-            </div>
-            <CardHeader>
-              <CardTitle>{station.chargingStation.name}</CardTitle>
-            </CardHeader>
-            <CardContent>
-              <div className="space-y-4">
-                <div>
-                  <p className="text-sm text-gray-500">Location</p>
-                  <p className="font-medium">
-                    {station.chargingStation.lat}, {station.chargingStation.lon}
-                  </p>
+          station.chargingStation ? (
+            <Card key={station.chargingStation.id} className="overflow-hidden">
+              <div className="relative h-48 w-full bg-gray-100 flex items-center justify-center">
+                <Image
+                  src="/charging-station.jpg"
+                  alt={station.chargingStation.name || "Charging Station"}
+                  fill
+                  className="object-cover"
+                  onError={(e) => {
+                    const target = e.target as HTMLImageElement
+                    target.style.display = 'none'
+                  }}
+                />
+                <div className="absolute inset-0 flex items-center justify-center">
+                  <Zap className="w-24 h-24 text-gray-400" />
                 </div>
-                
-                <div>
-                  <p className="text-sm text-gray-500">Availability</p>
-                  <div className="flex items-center justify-between">
+              </div>
+              <CardHeader>
+                <CardTitle>{station.chargingStation.name || "Unnamed Station"}</CardTitle>
+              </CardHeader>
+              <CardContent>
+                <div className="space-y-4">
+                  <div>
+                    <p className="text-sm text-gray-500">Location</p>
                     <p className="font-medium">
-                      {station.spots.filter(spot => spot.state === 'FREE').length} / {station.spots.length} spots
+                      {station.chargingStation.lat}, {station.chargingStation.lon}
                     </p>
-                    <span className={`px-2 py-1 rounded-full text-xs font-medium ${getAvailabilityColor(station)}`}>
-                      {station.spots.filter(spot => spot.state === 'FREE').length === 0 ? 'Full' : 'Available'}
-                    </span>
                   </div>
-                </div>
+                  
+                  {station.spots ? (
+                    <>
+                      <div>
+                        <p className="text-sm text-gray-500">Availability</p>
+                        <div className="flex items-center justify-between">
+                          <p className="font-medium">
+                            {station.spots.filter(spot => spot.state === 'FREE').length} / {station.spots.length} spots
+                          </p>
+                          <span className={`px-2 py-1 rounded-full text-xs font-medium ${getAvailabilityColor(station)}`}>
+                            {station.spots.filter(spot => spot.state === 'FREE').length === 0 ? 'Full' : 'Available'}
+                          </span>
+                        </div>
+                      </div>
 
-                <div>
-                  <p className="text-sm text-gray-500">Connector Types</p>
-                  <div className="flex flex-wrap gap-2 mt-1">
-                    {Array.from(new Set(station.spots.map(spot => spot.connectorType).filter(Boolean))).map(type => (
-                      <span key={type} className="px-2 py-1 bg-gray-100 rounded-full text-xs">
-                        {type}
-                      </span>
-                    ))}
-                  </div>
+                      <div>
+                        <p className="text-sm text-gray-500">Connector Types</p>
+                        <div className="flex flex-wrap gap-2 mt-1">
+                          {Array.from(new Set(station.spots.map(spot => spot.connectorType).filter(Boolean))).map(type => (
+                            <span key={type} className="px-2 py-1 bg-gray-100 rounded-full text-xs">
+                              {type}
+                            </span>
+                          ))}
+                        </div>
+                      </div>
+                    </>
+                  ) : (
+                    <p>No spot data available</p>
+                  )}
                 </div>
-              </div>
-            </CardContent>
-          </Card>
+              </CardContent>
+            </Card>
+          ) : null
         ))}
       </div>
     </div>
   )
-} 
+}
