--- green-ev/next.config.js
@@ -1,20 +1,0 @@
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-  webpack: (config) => {
-    config.module.rules.push({
-      test: /\.css$/,
-      use: [
-        'style-loader',
-        {
-          loader: 'css-loader',
-          options: {
-            importLoaders: 1,
-          },
-        },
-      ],
-    })
-    return config
-  },
-}
-
-module.exports = nextConfig
